Modifying Field3D iterator for single index
Iterating with a single index avoids calculating separate
(x,y,z) indices, and should enable better optimisation of loops.

Rather than allowing an (int) indexing, this wraps an integer
in a struct SingleIndexIterator<T> which prevents accidental indexing
of a 2D field with a 3D index (for example).

There are now two different ways to iterate over a (3D) field:

1. single index

    for(auto &i : f) {
      f[i] = ...
    }

  This is efficient, and iterates over the entire field. i is a SingleIndexIterator,
  which does not allow indexing of other field types, or stencil operations.

2. DataIterator indexing

    for(auto &i : f.region(RGN_ALL) {
      f[i] = ...
    }

  This uses a DataIterator, iterating over a specified region. This may be less
  efficient, but allows indexing of different field types, and stencil operations.