Remove interp_to(g_22) from Grad2_par2
This is the only differential operator that calls interp_to on the
metric coefficient. At the moment interp_to(Field2D) does nothing.
Removing this call so that behaviour does not change when making a start
on implementing interp_to for Field2D, until the metric coefficients are
sorted out properly for staggered grids.