Make BoutOutputs.evolvingVariables a public attribute
BoutOutputs.evolvingVariables() was only a getter function for a private
instance attribute. This is not good Python style:
https://www.python-course.eu/python3_properties.php
It is better to use a public instance attribute, then if getter/setter
type functionality is needed later, use @property and, e.g.,
@evolvingVariables.setter decorators. This allows for example checks to
be done when setting an attribute, but still allows the class user to
access like a variable instead of a function. For example:
  >> data = BoutOutputs()
  >> print(data.evolvingVariables)
     ['n', 'T']
instead of:
  >> data = BoutOutputs()
  >> print(data.evolvingVariables())
     ['n', 'T']