Check if phases have been calculated with member bools not static bool
Using static bool to check for cached phases means the check is global
so if two ShiftToFieldAligned objects are created the one which calls
'get...' second will fail. This is unlikely in a normal simulation but
might happen if there are multiple meshes, for example in tests or
post-processing using boutcore.